% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/content.R
\name{guardian_content}
\alias{guardian_content}
\title{Search Guardian Content}
\usage{
guardian_content(api_key, query, from = NULL, to = NULL, section = NULL,
  reference = NULL, reference_type = NULL, tags = NULL, rights = NULL,
  ids = NULL, production_office = NULL, page = NULL, page_size = 50,
  fields = NULL, ...)
}
\arguments{
\item{api_key}{A key to the Guardian API, which can be obtained
\href{http://open-platform.theguardian.com/access/}{here}.}

\item{query}{Your search query. This can contain operators (\code{sausage AND mash}) or
phrases (\code{"sausage & mash"}); by default, searches work as an OR, looking for
the presence of any one individual word in the query.}

\item{from}{the date to limit by. If set, the search will only cover data written after this date.
Optional (NULL) by default. Can be set in conjunction with \code{to}.}

\item{to}{Another date to limit by. If set, the search will only cover data written before this date.
Optional (NULL) by default. Can be set in conjunction with \code{from}.}

\item{section}{}

\item{reference}{}

\item{reference_type}{}

\item{tags}{}

\item{rights}{}

\item{ids}{}

\item{production_office}{}

\item{page}{}

\item{page_size}{}

\item{fields}{}

\item{...}{further arguments to pass to httr's \code{GET}.}
}
\description{
\code{guardian_content} lets you directly search
The Guardian for content, including support for conditional
searches, date-range based filtering, and section or tag
based filtering.
}

