% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editions.R
\name{guardian_editions}
\alias{guardian_editions}
\title{Search Guardian Editions}
\usage{
guardian_editions(api_key, query, ...)
}
\arguments{
\item{api_key}{A key to the Guardian API, which can be obtained
\href{http://open-platform.theguardian.com/access/}{here}.}

\item{query}{Your search query.}

\item{...}{further arguments to pass to httr's \code{GET}.}
}
\description{
\code{guardian_editions} lets you directly search
The Guardian for metadata about their editions - country-specific
(or international) standard releases for their website's content.
}
\examples{
\dontrun{
# Simple example
uk_edition_data <- guardian_editions("test", "uk")
}
}
\seealso{
\code{\link{guardian_sections}} for retrieving sections, another category
of content.
}

