% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sections.R
\name{guardian_sections}
\alias{guardian_sections}
\title{Search Guardian Sections}
\usage{
guardian_sections(api_key, query, ...)
}
\arguments{
\item{api_key}{A key to the Guardian API, which can be obtained
\href{http://open-platform.theguardian.com/access/}{here}.}

\item{query}{Your search query. This can contain operators (\code{sausage AND mash}) or
phrases (\code{"sausage & mash"}); by default, searches work as an OR, looking for
the presence of any one individual word in the query.}
}
\description{
\code{guardian_sections} lets you directly search
The Guardian for metadata about sections - individual categories of
content, such as world or US news.
}

