% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tags.R
\name{guardian_tags}
\alias{guardian_tags}
\title{Search Guardian Tags}
\usage{
guardian_tags(api_key, query, type = NULL, section = NULL,
  reference = NULL, reference_type = NULL, page = NULL, page_size = 50,
  ...)
}
\arguments{
\item{api_key}{A key to the Guardian API, which can be obtained
\href{http://open-platform.theguardian.com/access/}{here}.}

\item{query}{Your search query. This can contain operators (\code{sausage AND mash}) or
phrases (\code{"sausage & mash"}); by default, searches work as an OR, looking for
the presence of any one individual word in the query.}

\item{section}{the section, or sections, of \emph{The Guardian} that you want to limit the search to. Multiple
sections may be concatenated together using boolean operators; see \code{\link{rdian_and}} and \code{\link{rdian_or}}.}

\item{reference}{the references to limit the search to; only tags that include those references (and meet other
conditions) will be returned. Also accepts boolean operators.}

\item{reference_type}{the type of reference (such as \code{isbn}). Also accepts boolean operators.}
}
\description{
\code{guardian_tags} lets you directly search
The Guardian for their tags - metadata used to classify content.
}
\examples{
# Simple example
results <- guardian_tags("test", "green")
}
\seealso{
\code{\link{guardian_content}}.
}

