\name{rdian-package}
\alias{rdian-package}
\alias{rdian}
\docType{package}
\title{
\packageTitle{rdian}
}
\description{
\packageDescription{rdian}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rdian}
\packageIndices{rdian}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rdian}

Maintainer: \packageMaintainer{rdian}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
