% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdian.R
\docType{package}
\name{rdian}
\alias{rdian}
\alias{rdian-package}
\title{Modern API client for The Guardian}
\description{
This package provides a modern (httr/curl/jsonlite-based) API
client for The Guardian (https://guardian.co.uk), allowing useRs to retrieve
content and content metadata.
}

